Group 1: bootstrap labeling toolchain + dataset validation
Implements the data-acquisition pipeline that converts the
existing Mask R-CNN model's predictions into YOLO det labels,
plus the validation/visualization scaffolding to gate Phase 1.

New tools (run on dev PC, except where noted):
- tools/bootstrap_labeling.py
    Runs INSIDE .legacy-venv (Python 3.8 + TF 1.15 + mrcnn).
    Loads the legacy weights, infers on every image in the given
    directories, and writes one YOLO det .txt per image. Tracks
    per-class counts, confidence stats, and a `review_files` list
    of low-confidence images for prioritized Roboflow inspection.
- tools/validate_dataset.py
    Pre-training health check: cross-references images vs labels,
    parses every label line, flags out-of-range coordinates and
    classes, computes class imbalance and empty-label ratio.
- tools/visualize_labels.py
    Renders bbox overlays for human review, with a --review-list
    flag to render only the bootstrap-flagged subset.
- training/data.yaml
    YOLO dataset definition. Source of truth for class taxonomy
    at training time, kept in sync with inference/types.py.

Tests (43 passing in main venv):
- tests/unit/test_types.py — class taxonomy invariants, Detection
  immutability + JSON shape, letterbox round-trip math.
- tests/unit/test_bootstrap_labeling.py — coordinate conversion
  edge cases (clamping, rectangular images), image iteration
  contract.
- tests/unit/test_validate_dataset.py — every invalid-line case,
  cross-reference logic, class imbalance calculation.

Tested with ruff (0 errors) and pytest (43/43 passed).

+++ inference/types.py
@@ -7,7 +7,7 @@ only on `numpy` and stdlib, no GStreamer or hailo_platform.
 
 from __future__ import annotations
 
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from pathlib import Path
 from typing import Final
 
@@ -121,7 +121,7 @@ def letterbox_size(orig_shape: tuple[int, int], target: int) -> tuple[int, int, 
     """
     h, w = orig_shape
     scale = min(target / h, target / w)
-    return int(round(h * scale)), int(round(w * scale)), scale
+    return round(h * scale), round(w * scale), scale
 
 
 def unletterbox_bbox(
